Parameter Make.1-B

include Mirage_block.S

type error = private [>

| Mirage_device.error

]

val pp_error : error Fmt.t

type write_error = private [>

| `Disconnected

| `Is_read_only

| `Unimplemented

]

val pp_write_error : write_error Fmt.t

type t

val disconnect : t -> unit Lwt.t

val get_info : t -> Mirage_block.info Lwt.t

val read : t -> int64 -> Cstruct.t list -> (unit, error) Stdlib.result Lwt.t

val write : t -> int64 -> Cstruct.t list -> (unit, write_error) Stdlib.result Lwt.t

val resize : t -> int64 -> (unit, write_error) Result.result Lwt.t

Resize the file to the given number of sectors.

val flush : t -> (unit, write_error) Result.result Lwt.t

flush t flushes any buffers, if the file has been opened in buffered mode