Module Qcow_cluster_map.Available

Clusters which are available for reallocation

val get : t -> Qcow_types.Cluster.IntervalSet.t

get t query the current contents of the set

val remove : t -> Qcow_types.Cluster.IntervalSet.t -> unit

remove t less removes less from the set

val mem : t -> Qcow_types.Cluster.t -> bool

mem t cluster is true if cluster is in t