Module Lwt_error.List

val map_p : ('a -> ('b, 'error) Result.result Lwt.t) -> 'a list -> ('b list, 'error) Result.result Lwt.t

map_p f xs computes f x where x \in xs concurrently and returns a list of successful results or the first error encountered. All threads will have terminated by the time the function returns.