Module Qcow_header.Feature

type ty = [

| `Incompatible

| `Compatible

| `Autoclear

]

type feature = [

| `Corrupt

| `Dirty

| `Lazy_refcounts

| `Unknown of string

]

type t = {

ty : ty;

bit : int;

feature : feature;

}

val understood : t list

The features understood by this implementation

include Qcow_s.SERIALISABLE with type t := t

type t

Instances of this type can be read and written

val sizeof : t -> int

The size of a buffer needed to hold t

val read : Cstruct.t -> (t * Cstruct.t, [ `Msg of string ]) Result.result

Read a t from the given buffer and return it, along with the unused remainder of the buffer. If the buffer cannot be parsed then return an error.

val write : t -> Cstruct.t -> (Cstruct.t, [ `Msg of string ]) Result.result

Write a t into the given buffer. If the buffer is too small, then return an error. Return the unused remainder of the buffer.