Module Qcow_int64

Parsers and printers for types used in qcow2 fields

include module type of Stdlib.Int64

val zero : int64

val one : int64

val minus_one : int64

val neg : int64 -> int64

val add : int64 -> int64 -> int64

val sub : int64 -> int64 -> int64

val mul : int64 -> int64 -> int64

val div : int64 -> int64 -> int64

val unsigned_div : int64 -> int64 -> int64

val rem : int64 -> int64 -> int64

val unsigned_rem : int64 -> int64 -> int64

val succ : int64 -> int64

val pred : int64 -> int64

val abs : int64 -> int64

val max_int : int64

val min_int : int64

val logand : int64 -> int64 -> int64

val logor : int64 -> int64 -> int64

val logxor : int64 -> int64 -> int64

val lognot : int64 -> int64

val shift_left : int64 -> int -> int64

val shift_right : int64 -> int -> int64

val shift_right_logical : int64 -> int -> int64

val of_int : int -> int64

val to_int : int64 -> int

val unsigned_to_int : int64 -> int option

val of_float : float -> int64

val to_float : int64 -> float

val of_int32 : int32 -> int64

val to_int32 : int64 -> int32

val of_nativeint : nativeint -> int64

val to_nativeint : int64 -> nativeint

val of_string : string -> int64

val of_string_opt : string -> int64 option

val to_string : int64 -> string

val bits_of_float : float -> int64

val float_of_bits : int64 -> float

type t = int64

val compare : t -> t -> int

val unsigned_compare : t -> t -> int

val equal : t -> t -> bool

val format : string -> int64 -> string

val t_of_sexp : Sexplib.Sexp.t -> t

val sexp_of_t : t -> Sexplib.Sexp.t

val of_int64 : int64 -> t

val to_int64 : t -> int64

val round_up : int64 -> int64 -> int64

round_up value to rounds value to the next multiple of to

val round_down : int64 -> int64 -> int64

round_down value to rounds value down to the multiple of to

module IntervalSet : Qcow_s.INTERVAL_SET with type elt = t

module Map : Stdlib.Map.S with type Map.key = t

include Qcow_s.SERIALISABLE with type t := t

type t

Instances of this type can be read and written

val sizeof : t -> int

The size of a buffer needed to hold t

val read : Cstruct.t -> (t * Cstruct.t, [ `Msg of string ]) Result.result

Read a t from the given buffer and return it, along with the unused remainder of the buffer. If the buffer cannot be parsed then return an error.

val write : t -> Cstruct.t -> (Cstruct.t, [ `Msg of string ]) Result.result

Write a t into the given buffer. If the buffer is too small, then return an error. Return the unused remainder of the buffer.