Module Qcow_metadata.Physical

type t

A cluster full of 64 bit cluster pointers

val of_contents : contents -> t

Interpret the given cluster as a cluster of 64 bit pointers

val get : t -> int -> Qcow_physical.t

get t n return the nth physical address within t

val set : t -> int -> Qcow_physical.t -> unit

set t n v set the nth physical address within t to v

val len : t -> int

len t returns the number of physical addresses within t