Module Qcow_metadata.Refcounts

type t

A cluster full of 16bit refcounts

val of_contents : contents -> t

Interpret the given cluster as a refcount cluster

val get : t -> int -> int

get t n return the nth refcount within t

val set : t -> int -> int -> unit

set t n v set the nth refcount within t to v