Module Qcow_rwlock

type t

A lock which permits multiple concurrent threads to acquire it for reading but demands exclusivity for writing

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

type ts = t list

val sexp_of_ts : ts -> Ppx_sexp_conv_lib.Sexp.t

val make : (unit -> string) -> t

make describe_fn creates a new lock, where describe_fn () returns a human-readable description string suitable for debug output.

type lock

A value which represents holding a lock

val unlock : lock -> unit

unlock locked releases the lock associated with locked

module Client : sig ... end

module Read : sig ... end

module Write : sig ... end

module Debug : sig ... end