Module Qcow_types

Parsers and printers for types used in qcow2 fields

val big_enough_for : string -> Cstruct.t -> int -> unit Qcow_error.t

big_enough_for name buf length returns an error with a log message if buffer buf is smaller than length. The name will be included in the error message.

module Int8 : sig ... end

module Int16 : sig ... end

module Int32 : sig ... end

module Int64 : module type of Qcow_int64

module Int : module type of Qcow_int

module Cluster : sig ... end